Moved restoreRadios function to funcs folder
